🎯 Refactor code structure
